+++ src/__tests__/intitusjonsopphold/Tabell.test.tsx
@@ -2,7 +2,7 @@ import "@testing-library/jest-dom";
 import renderer from "react-test-renderer";
 import { IntlProvider } from "react-intl";
 import nbMessages from "text/nb";
-import Tabell from "../../pages/institusjonsopphold/historikk/Tabell";
+import InstHistorikkView from "../../pages/institusjonsopphold/historikk/InstHistorikkView";
 import instInfo from "../../clients/apiMock/app/fetch/inst-info.json";
 
 jest.mock("react-router-dom", () => ({
@@ -14,12 +14,12 @@ jest.mock("react-router-dom", () => ({
 
 jest.mock("nav-frontend-hjelpetekst");
 
-describe("Tabell", () => {
+describe("InstHistorikkView", () => {
   it("renders correctly", () => {
     const tree = renderer
       .create(
         <IntlProvider locale={"nb"} messages={nbMessages}>
-          <Tabell instInfo={instInfo} />
+          <InstHistorikkView instInfo={instInfo} />
         </IntlProvider>
       )
       .toJSON();
